Extract function

+++ src/app/layer-managers/markers-layer-manager.ts
@@ -200,28 +200,17 @@ class MarkersLayerManager {
     const deleteTitleText = this.translateService.instant(
       'markers.deleteConfirmHeader',
     );
-    const cancelText = this.translateService.instant('general.cancel');
     const deleteText = this.translateService.instant('general.delete');
 
-    const alert = await this.alertController.create({
-      header: deleteTitleText,
-      subHeader: marker.name,
-      buttons: [
-        {
-          text: cancelText,
-          role: 'cancel',
-        },
-        {
-          text: deleteText,
-          handler: async () => {
-            await this.markersService.delete(marker.id);
-            await alert.dismiss();
-            await this.reloadAllMarkers();
-          },
-        },
-      ],
-    });
-
-    await alert.present();
+    await this.alertController.confirm(
+      deleteTitleText,
+      marker.name,
+      deleteText,
+      async (alert) => {
+        await this.markersService.delete(marker.id);
+        await alert.dismiss();
+        await this.reloadAllMarkers();
+      },
+    );
   }
 }
